requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-discover-runner==1.0
django-role-permissions==3.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
Pillow==9.4.0
pkg_resources==0.0.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
